requirements.txt
Pillow
reportlab
bs4
pyPDF2
psutil
openai
packaging